[feat] MainCard 내 useEffect 부분을 상위 useLoaData 내에 이동
